Dev-tools autocomplete bool/filter with array.
The Dev-tools auto-complete for filter should be identical to the must, should, and must_not sections.  They take an array of sub-queries.

Note that there has not been any syntactic distinction between queries and filters since version 2.0 of Elasticsearch and so filter, must, should, and must_not are supposed to all have the same "__scope_link".

Git issue: #20256